Add leaderelectin options to gitjob controller
LeaseDuration, RenewDeadline and RetryPeriod will be set for the gitjob
controller and their values determined by environment variables.
