+++ internal/cmd/controller/gitops/operator.go
@@ -10,6 +10,7 @@ import (
 
 	command "github.com/rancher/fleet/internal/cmd"
 	"github.com/rancher/fleet/internal/cmd/controller/gitops/reconciler"
+	"github.com/rancher/fleet/internal/cmd/controller/options"
 	"github.com/rancher/fleet/internal/metrics"
 	fleet "github.com/rancher/fleet/pkg/apis/fleet.cattle.io/v1alpha1"
 	"github.com/rancher/fleet/pkg/git/poll"
@@ -86,12 +87,20 @@ func (g *GitOperator) Run(cmd *cobra.Command, args []string) error {
 
 	namespace := g.Namespace
 
+	leaderOpts, err := options.NewLeaderElectionOptions()
+	if err != nil {
+		return err
+	}
+
 	mgr, err := ctrl.NewManager(ctrl.GetConfigOrDie(), ctrl.Options{
 		Scheme:                  scheme,
 		Metrics:                 g.setupMetrics(),
 		LeaderElection:          g.EnableLeaderElection,
 		LeaderElectionID:        "gitjob-leader",
 		LeaderElectionNamespace: namespace,
+		LeaseDuration:           leaderOpts.LeaseDuration,
+		RenewDeadline:           leaderOpts.RenewDeadline,
+		RetryPeriod:             leaderOpts.RetryPeriod,
 	})
 
 	if err != nil {
